backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
sqlalchemy==2.0.49
greenlet==3.4.0
aiosqlite==0.22.1
asyncpg==0.30.0
alembic==1.18.4
redis==7.4.0
python-jose[cryptography]==3.5.0
python-multipart==0.0.26
websockets==13.1
